ref(persister): collapse wallet subdir to persister
- move wallet subdir into the persister module
to simplify the structure as wallet was confusing
